+++ lib/WorseReflection/Core/Reflection/OldCollection/ReflectionConstantCollection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\WorseReflection\Core\Reflection\OldCollection;
+namespace Phpactor\WorseReflection\Core\Reflection\Collection;
 
 use Phpactor\WorseReflection\Core\Reflection\ReflectionConstant;
